Introduce bitstream.Walker
bitstream.Reader is a low-level API that handles parsing bistream
values one at a time, but it is unaware of a bitstream structure;
it is up to the caller to use all API functions in the correct order,
and to validate things like abbreviation definitions after sequential
values have been read. It stores no context other than stream offset.

bitstream.Walker is an attempt to provide a safe way to iterate all
blocks and records in a bitstream, while still not being opinionated
or informed about the structure expected by an application.

Some assumptions made:

- All "codes" (block IDs, record IDs) are 32-bit.
- All record values can be up to 64-bit.
- An array in an abbreviation does not imply that the application
  expects an array-typed value -- it might just be a compact way to
  repeat the same encoding for several operands. So, the walker keeps
  a state machine to track progress through a terminating array, same
  as it does for positional operands.
- A blob terminating an abbreviation is such a special case that it is
  considered an expected type. Therefore, a walker does not iterate
  through blob values, but only has the ability to return a slice
  of its data. This simplified walker bookkeeping. Also, an application
  must explicitly request a blob from the record, and it must do so
  after reading (or skipping) all previous operands.
- It is never safe to skip blocks. The bitstream spec places no
  restrictions on where a BLOCKINFO can appear. Because BLOCKINFO can
  define encodings for blocks at any level, and the order of
  abbreviation definitions determines their IDs, skipping any block
  could result in corrupt decoding.

Walker's core method is `next()`, which returns block enter/exit
events and the presence of a record. Records can be parsed with
specialized methods (which are only legal when a record is
encountered), but records are safe to skip with another call to
`next()`, regardless of whether they have been fully read.

Walker hides details of how a record is encoded, because applications
should not expect records to be encoded in any specific way. The
exception to this rule is blobs.

Applications can and do sometimes expect the remaining sequential
values in a record to represent a string or other array. Because of
arrays and the way they are encoded, the number of remaining record
values cannot always be determined at the current cursor position.
For now, Walker only exposes a way to collect these values with an
allocator, rather than a `read(buf: []const T)` API.

At initialization, Walker takes a slice of the bitstream data instead
of an io.Reader. This is for a few reasons:

- While the ability to operate on an arbitrary stream is nice, in
  reality all data will come from files on disk or an in-memory
  buffer, not e.g. a network socket.
- Bitstream files are typically small, in the KB to MB range, so it is
  reasonable to expect them to be stored in memory.
- Not having a reference to the soure slice would mean it could not
  return a slice for blobs. It could return a tuple with start/end
  offsets, but that feels hacky and not ergonomic to use for the
  caller.
- The internal BitReader state is highly dependent on the state of
  the Reader it wraps. If the caller created the Reader, there is a
  greater risk that they might use it themselves, which would corrupt
  the BitReader state.

The need to have both a BitReader and FixedBufferStream stored in the
Walker struct led to a less-than-ideal `init()` API because of the
interdependent pointer fields. But it's still better than a constructor
which allocates on the heap.