Rewrite parser to be not-generally recursive
LLVM Bitcode files are pretty fixed in structure; they aren't really
recursive. So I changed the parser to be a little more rigid and
straightforward. As a result, it uses a little bit less comptime
trickery.

The benefit of this is it's easier to pass extra arguments to a
block parsing helper when context is significant. Namely, parsing
a "constants" block will put the data in either the module or a
specific function, depending where it is found.

I've also removed some of the forgiveness when an unknown block
is encountered, but might revisit that later.

Also added record codes for constant defs and fixed a small
bug in walker.