Delete old parser and fix leaks
I'm not super satisfied returning the arena in the Result - I would
prefer to define `res.deinit(allocator)`. There's no quick and easy
way to do that because the Bitcode is a complicated structure suited
well to an arena allocator, but I would rather not have `parse()` take
an arena. I might revisit this design later.